refactor(core,query): relocate makeRouteSafe to @orval/core
Move makeRouteSafe and wrapRouteParameters from @orval/query/utils into
@orval/core/getters/route so non-query packages can reuse them without
introducing a reverse dependency on @orval/query. Update query/src/client.ts
to import makeRouteSafe from @orval/core and move the related unit tests
to core/getters/route.test.ts. Behavior is unchanged.

This prepares for the follow-up fix that restores urlEncodeParameters
support in @orval/fetch (#3343).
